requirements-freeze.txt
filelock==3.18.0
fsspec==2025.3.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
pandas==2.2.3
pillow==11.1.0
py4j==0.10.9.7
pyarrow==19.0.1
pyspark==3.5.5
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.13.0
tzdata==2025.2
